+++ app/controllers/api/v1/weapons_controller.rb
@@ -7,19 +7,25 @@ module Api
       before_action :set_weapon, only: %i[show update destroy]
 
       def show
+        authorize @weapon
         render json: @weapon
       end
 
       def create
         @character = Character.find(params[:character_id])
-        render json: Weapon.create(weapon_params)
+        @weapon = Weapon.create(weapon_params)
+        @weapon.character = @character
+        authorize @weapon
+        render json: @weapon
       end
 
       def destroy
+        authorize @weapon
         render json: @weapon.destroy
       end
 
       def update
+        authorize @weapon
         @weapon.update_attributes(weapon_params)
         render json: @weapon
       end
